requirements.txt
selenium==3.141.0
python-dotenv==0.19.0
yagmail==0.14.256